fix: polyfill Data3DTexture
